Break CLI into multiple subcommands
The CLI had several options that were mutually exclusive, as well as options that behaved differently depending on some of those exclusive options (for example, `--output` behaved differently depending on whether `--generate_docs` was set). This breaks the CLI up into multiple subcommands, both to clarify behavior for users and to make the implementation a bit more organized.

The new commands are `data` (the "main" command that most options used to apply to), `list`, and `generate-docs`. I've also tried to rework the help text here to be clearer and more informative. Unsupported options that are placeholders for future behavior now throw errors when used rather than silently failing.

Finally, options have all been updated to use dash-separated words (there was previously a mix of underscores and dashes).