Move documentation defaults up to CLI
The default for `--output-dir` for docs generation is buried pretty deep in `DocumentationFactory.writeDocumentationForSchema` logic, when it could have been handled more idiomatically in the function arguments or in the CLI code, where it's a little more relevant (since it's working-directory relative, rather than source-file-relative). This makes the default in the function more idiomatic by using default arguments and more generic, then moves the actual, meaningful default up to the CLI where it can be documented more clearly.