Add a PR check to verify that docs are up-to-date
Documentation that is generated based on schema files is currently stored in the repository. However, sometimes contributors update the schema files without including the generated docs. Not only does this result in out of date docs, but when someone else makes a change later and *does* generate new docs, their patches are full of erroneous docs changes, making them harder to evaluate.

This change adds a workflow job to check whether there are uncommitted docs changes on every PR.