3869: Accept Internatinal Phone Number
We currently have the libphonenumber library in the application,
which comes from Google. This should be able to handle international
phone numbers. We should use the library to verify that it is a valid
phone number, and if it is, then we can accept it. If it is NOT a
North American phone number (therefore Canada, Mexico, or US), we
should not fill in PID.13.6 and PID.13.7, but instead just fill in
PID.13.1 and PID.13.12. We SHOULD always fill in PID.13.4,
which is country code.