Update template for linked issues (#5423)
Updates the section for issues fixed by the PR to demonstrate the format for linking issues in pull requests. GitHub markdown will hyperlink an issues written as `#1234` but to formally link it to the PR you need to preface it with a supported keyword. 

See https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword