Add ActiveValues concept to report Options
When submitting a deprecated or invalid option, the error/warning message returns a list of all the options. Unfortunately all the options includes the deprecated options as well so you get the confusing message of "x option has been deprecated. Valid options: a, b, c, d, x" This commit creates an active values concept and uses it for all error messages