Experience 7606: Wrap console.warn, console.error to report to AI
This changeset carries over [the SimpleReport TelemetryService logic](https://github.com/CDCgov/prime-simplereport/blob/main/frontend/src/app/TelemetryService.ts) that we want to use here.  I started refactoring it, but I decided to just copy it over verbatim for consistency, and we can make modifications to it in the future if we need.  However, one thing that could help here could be to add a dev-only AI environment so that we can always initialize the ApplicationsInsights instance.  That'd obviate the need for a lazy initialization like we're doing with `ai.initialize()`, and we could test that events are actually getting passed to the proper channels locally prior to deplying to staging or demo environments.  Something we can discuss in the future, though 🤷