Experience-7607: Capture FE exceptions in Azure
This changeset reverts the revert in https://github.com/CDCgov/prime-reportstream/pull/7740.  There are two main changes from that revert:

1) Always pass an instance of Error to `trackException`
Originally, we were passing in either an Error or a string depending on the arguments `console.error` was called with.  However, `exception` should be an instance of [Error or IAutoExceptionTelemetry](https://github.com/microsoft/ApplicationInsights-JS/blob/master/shared/AppInsightsCommon/src/Interfaces/IExceptionTelemetry.ts#L31) or we'll end up with a `not_specified` for the exception message.

2) Only tracking error-level messages
This is a deviation from SimpleReport so I'm happy to discuss and update if needed, but my feeling is that we don't have to track warnings just yet until we have a proper funnel or response to unexpected behaviors.  Errors are much more significant to track, and while they -- like warnings -- can be triggered from anywhere (by third-party scripts, by browsers themselves, et cetera), they're generally more actionable or intentional than warnings.

Also included:
- Some minor cleanup
- Rewritten tests + new ones to check for the changes