Experience-7886: Fix memberships.has in admin check
This changeset addresses a bug wherein the `memberships.has` check for an admin status fails because `memberships` is not a Map.  This requires users to clear their local storage in order to continue using the site.

Since we no longer have the ability to mock a user's member type, we can just check if their active membership is an admin member type instead.