Experience: add template for frontend-react proposals, add number to existing proposal
This changeset starts to consolidate the ADR/proposal process in frontend-react, and it adds a template that we can use for ADRs going forward.  This is to provide historical context over what architectural options have already been decided on/against so that our team can be better informed should relevant discussions be raised again in the future.