Experience-7867: Add local Cypress UX checks (#7913)
This changeset includes a basic scaffolding of running Cypress tests locally.  It's far from complete coverage (whatever that means), but it covers basic navigation and includes some slightly more complicated things like determining what to check for by making API calls prior to rendering.  At a high level, there are a few concepts introduced, like extensible page objects and commands.

There are still lots of things we can do to build on this, but my immediate focus after this will be on:
- Running locally against other environments (demo1/2/3, staging, production)
- Cleaning up environment variable schtuff
- MFA