Experience-8108: Remove upload-to-send functionality (#8113)
We got the go-ahead from O+O to remove the upload-to-send functionality since that's been moved over to SimpleReport so that's what this changeset is doing!  It'll make the public validator work a bit easier too!

Note that this is strictly removing the _front-end_ logic for upload-to-send so as to make the public validator work slightly easier for us.  Nothing on the API is changing as a part of this changeset.