Experience-7773: Add schema dropdown on validator page
This changeset adds a dropdown to the validator page to allow the user to select between schema options to validate against.  To accommodate this, I've added a `useSenderSchemaOptions` hook that always returns the three "master" schemas" in addition to the designated Sender schema if the current user is logged is as a Sender.

This also includes a decent amount of cleanup/rewrites because the logic in useFileHandler, FileHandler, and FileHandlerForm is oddly coupled together.  Will comment below!

To generate test files to validate against different schemas, you can use the CLI tool:
```
./prime data --input-fake <item_count> --input-schema <schema> --output <output_destination>
```

like:

```
./prime data --input-fake 10 --input-schema hl7/hcintegrations-covid-19 --output ~/Desktop/whatever.hl7
```

This seems to generate lots of duplicate columns so those have to be removed or validation will complain about those, but this is a starting point for test data!