Experience-8161: Remove makeOktaHook logic (#8323)
This changeset removes the makeOktaHook logic that's fed into SessionProvider and a few test utilities.  There's a temporary workaround to fall back on an empty object in SessionProvider from the `useOktaAuth` call -- this is to account for components being rendered outside of Okta's Security Provider, in which case `useOktaAuth` calls return `null`.  This will be fixed as part of [#8292](https://app.zenhub.com/workspaces/experience-607d9d5e68b95200150fec37/issues/gh/cdcgov/prime-reportstream/8292); it'd just be easier to remove this logic beforehand.

Ultimately, there's no reason for us to have to pass a stand-in for `useOktaAuth` as a prop when we could just mock it if we really need to.