Experience-8026: Widen grid container
This changeset updates the follow pages' containers to be widescreen (up to 1400px) since they're the most data-heavy:
- Daily Data
- Submissions
- Message Tracker

The Public Validator page will continue to be on 1024px to align with the rest of the public pages.

Additionally, there's some cleanup regarding containers in the codebase across several different pages, and I've updated `grid-container` modifier classes to use the GridContainer component instead.  There are a lot of `grid-container` modifier classes used around the codebase, and it really shouldn't be used this much; it should be used at the highest ancestor node of common width, and all descendants should naturally expand to their own widths inside of that.  The grid container sets a maximum width, yes, but it also includes padding to account for grid gutters -- if this is used too frequently, there'll be a lot of horizontal jaggedness across the website.

**tl;dr In most cases, a grid container should be used at the top of the page and nowhere else.**  There are obviously use cases for nested grid containers, but they should be rare and should be brought up by design rather than used anywhere we need a maximum width -- especially not in components that should be able to be rendered agnostically to the wrapping context.