Experience-8549: Fix incorrect setting type history (#8550)
This changeset updates the setting revision query key to be cached on the Organization and setting type (Sender|Receiver).  Originally, it was being cached on the strings "org" and "settingType" verbatim, which means that all revision histories would get the same response that's in the query cache.  This just updates the query key to be more dynamic.