Experience 8616: Relax Organization/service name validation (#8760)
This changeset relaxes the Organization/service name validation to allow letters, numbers, hyphens, and underscores, and it removes the reserved word restrictions we had previously.  This was causing issues with current Sender/Receiver settings, and as far as I know, we haven't run into issues with them in production anyway.  It'd be best to run this validation on the back end instead of the front end since the user can make the PUT request themselves with invalid names, but this is just a little extra check/helpful reminder to admins to stay in line with our naming conventions.  Tests have been updated as well.